Removed models and refactored all models to separate files and imports
