+++ src/database/models/mod.rs
@@ -1,0 +1,6 @@
+pub mod auth;
+pub mod collection;
+pub mod enums;
+pub mod notifications;
+pub mod object;
+pub mod object_group;
